Initial work for deprecation of legacy APIs
