+++ src/api/Tools.ts
@@ -463,4 +463,13 @@ export namespace Tools {
     }
     return 0;
   }
+
+  export function fileToPath(file: string | vscode.Uri): string {
+    if (typeof file === "string") {
+      return Tools.fixWindowsPath(file);
+    }
+    else {
+      return file.fsPath;
+    }
+  }
 }
