+++ src/api/debug/certificates.ts
@@ -77,7 +77,7 @@ async function getExtFileContent(hostInfo: HostInfo) {
  * @param imported if defined, gives the location and password of a local or remote (i.e. on the IFS) service certificate to import
  */
 export async function setup(connection: IBMi, imported?: ImportedCertificate) {
-  if (!(await connection.checkUserSpecialAuthorities(["*ALLOBJ"])).valid) {
+  if (!(await connection.getContent().checkUserSpecialAuthorities(["*ALLOBJ"])).valid) {
     throw new Error(`User ${connection.currentUser} doesn't have *ALLOBJ special authority`);
   }
   await vscode.window.withProgress({ title: "Setup debug service", location: vscode.ProgressLocation.Window }, async (task) => {
@@ -91,7 +91,7 @@ export async function setup(connection: IBMi, imported?: ImportedCertificate) {
       }
     }
 
-    const debugConfig = await new DebugConfiguration().load();
+    const debugConfig = await new DebugConfiguration(connection).load();
 
     const certificatePath = debugConfig.getRemoteServiceCertificatePath();
     const directory = dirname(certificatePath);
@@ -109,7 +109,7 @@ export async function setup(connection: IBMi, imported?: ImportedCertificate) {
       password = imported.password;
       if (imported.localFile) {
         setProgress("importing local certificate");
-        await connection.uploadFiles([{ local: imported.localFile, remote: debugConfig.getRemoteServiceCertificatePath() }]);
+        await connection.getContent().uploadFiles([{ local: imported.localFile, remote: debugConfig.getRemoteServiceCertificatePath() }]);
       }
       else if (imported.remoteFile) {
         setProgress("importing remote certificate");
@@ -214,9 +214,10 @@ export async function debugKeyFileExists(connection: IBMi, debugConfig: DebugCon
 }
 
 export async function remoteCertificatesExists(debugConfig?: DebugConfiguration) {
-  const content = instance.getContent();
-  if (content) {
-    debugConfig = debugConfig || await new DebugConfiguration().load();
+  const connection = instance.getConnection();
+  if (connection) {
+    const content = connection.getContent();
+    debugConfig = debugConfig || await new DebugConfiguration(connection).load();
     return await content.testStreamFile(debugConfig.getRemoteServiceCertificatePath(), "f") && await content.testStreamFile(debugConfig.getRemoteClientCertificatePath(), "f");
   }
   else {
@@ -225,13 +226,8 @@ export async function remoteCertificatesExists(debugConfig?: DebugConfiguration)
 }
 
 export async function downloadClientCert(connection: IBMi) {
-  const content = instance.getContent();
-  if (content) {
-    await content.downloadStreamfileRaw((await new DebugConfiguration().load()).getRemoteClientCertificatePath(), getLocalCertPath(connection));
-  }
-  else {
-    throw new Error("Not connected to an IBM i");
-  }
+  const content = connection.getContent();
+  await content.downloadStreamfileRaw((await new DebugConfiguration(connection).load()).getRemoteClientCertificatePath(), getLocalCertPath(connection));
 }
 
 export function getLocalCertPath(connection: IBMi) {
@@ -242,7 +238,7 @@ export function getLocalCertPath(connection: IBMi) {
 export async function checkClientCertificate(connection: IBMi, debugConfig?: DebugConfiguration) {
   const locaCertificatePath = getLocalCertPath(connection);
   if (existsSync(locaCertificatePath)) {
-    debugConfig = debugConfig || await new DebugConfiguration().load();
+    debugConfig = debugConfig || await new DebugConfiguration(connection).load();
     const remote = (await connection.sendCommand({ command: `cat ${debugConfig.getRemoteClientCertificatePath()}` }));
     if (!remote.code) {
       const localCertificate = readFileSync(locaCertificatePath).toString("utf-8");
@@ -263,7 +259,7 @@ export async function sanityCheck(connection: IBMi, content: IBMiContent) {
   //Since Code for IBM i v2.10.0, the debug configuration is managed from the debug service .env file
   //The encryption key is backed up since it's destroyed every time the service starts up
   //The remote certificate is only valid if the client certificate is found too
-  const debugConfig = await new DebugConfiguration().load();
+  const debugConfig = await new DebugConfiguration(connection).load();
 
   //Check if java home needs to be updated if the service got updated (e.g: v1 uses Java 8 and v2 uses Java 11)
   const javaHome = debugConfig.get("JAVA_HOME");
@@ -293,7 +289,7 @@ export async function sanityCheck(connection: IBMi, content: IBMiContent) {
       })
     }
   }
-  else if ((await connection.checkUserSpecialAuthorities(["*ALLOBJ"])).valid) {
+  else if ((await connection.getContent().checkUserSpecialAuthorities(["*ALLOBJ"])).valid) {
     try {
       if (legacyCertExists && !remoteCertExists) {
         //import legacy
