+++ src/api/debug/index.ts
@@ -259,7 +259,7 @@ export async function initialize(context: ExtensionContext) {
         if (doSetup) {
           Tools.withContext("code-for-ibmi:debugWorking", async () => {
             if (!(await server.getDebugServiceJob()) || await server.stopService(connection)) {
-              const debugConfig = await new DebugConfiguration().load();
+              const debugConfig = await new DebugConfiguration(connection).load();
               const clearResult = await connection.sendCommand({ command: `rm -f ${debugConfig.getRemoteServiceCertificatePath()} ${debugConfig.getRemoteClientCertificatePath()}` });
               if (clearResult.code) {
                 vscode.window.showErrorMessage(`Failed to clear existing certificate: ${clearResult.stderr}`);
@@ -278,12 +278,12 @@ export async function initialize(context: ExtensionContext) {
     vscode.commands.registerCommand(`code-for-ibmi.debug.setup.remote`, (doSetup?: 'Generate' | 'Import') =>
       Tools.withContext("code-for-ibmi:debugWorking", async () => {
         const connection = instance.getConnection();
-        const content = instance.getContent();
-        if (connection && content) {
+        if (connection) {
+          const content = connection.getContent();
           const ptfInstalled = server.debugPTFInstalled();
 
           if (ptfInstalled) {
-            const debugConfig = await new DebugConfiguration().load()
+            const debugConfig = await new DebugConfiguration(connection).load()
             const remoteCertsExists = await certificates.remoteCertificatesExists(debugConfig);
             if (remoteCertsExists && await certificates.debugKeyFileExists(connection, debugConfig) && debugConfig.getCode4iDebug()) {
               await certificates.downloadClientCert(connection);
