+++ src/testing/action.ts
@@ -63,7 +63,8 @@ export const ActionSuite: TestSuite = {
   tests: [
     {
       name: `Variable expansion test`, test: async () => {
-        const result = await CompileTools.runCommand(instance, {
+        const connection = instance.getConnection()!;
+        const result = await CompileTools.runCommand(connection, {
           command: 'echo "&CURLIB &MYTEXT"',
           env: { '&MYTEXT': `&BRANCHLIB &BRANCH`, '&BRANCHLIB': 'MYLIB', '&BRANCH': 'my/lib' },
           environment: `pase`
