+++ src/webviews/login/index.ts
@@ -2,7 +2,7 @@ import vscode, { l10n, ThemeIcon } from "vscode";
 import { ConnectionConfiguration, ConnectionManager } from "../../api/Configuration";
 import { CustomUI, Section } from "../../api/CustomUI";
 import IBMi from "../../api/IBMi";
-import { disconnect, instance } from "../../instantiate";
+import { safeDisconnect, instance } from "../../instantiate";
 import { ConnectionData } from '../../typings';
 
 type NewLoginSettings = ConnectionData & {
@@ -21,7 +21,7 @@ export class Login {
   static async show(context: vscode.ExtensionContext) {
     const connection = instance.getConnection();
     if (connection) {
-      if (!disconnect()) return;
+      if (!safeDisconnect()) return;
     }
 
     const connectionTab = new Section()
@@ -115,7 +115,7 @@ export class Login {
                     }
 
                   } else {
-                    vscode.window.showErrorMessage(`Not connected to ${data.host}! ${connected.error.message || connected.error}`);
+                    vscode.window.showErrorMessage(`Not connected to ${data.host}! ${connected.error}`);
                   }
                 } catch (e) {
                   vscode.window.showErrorMessage(`Error connecting to ${data.host}! ${e}`);
@@ -145,7 +145,7 @@ export class Login {
       // If the user is already connected and trying to connect to a different system, disconnect them first
       if (name !== existingConnection.currentConnectionName) {
         vscode.window.showInformationMessage(`Disconnecting from ${existingConnection.currentHost}.`);
-        if (!await disconnect()) return false;
+        if (!await safeDisconnect()) return false;
       }
     }
 
@@ -175,7 +175,7 @@ export class Login {
         if (connected.success) {
           vscode.window.showInformationMessage(`Connected to ${connectionConfig.host}!`);
         } else {
-          vscode.window.showErrorMessage(`Not connected to ${connectionConfig.host}! ${connected.error.message || connected.error}`);
+          vscode.window.showErrorMessage(`Not connected to ${connectionConfig.host}! ${connected.error}`);
         }
 
         return true;
